refactor: split perf into multiple files
