+++ perf/Cargo.toml
@@ -25,7 +25,6 @@ serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0.94"
 sp1-sdk = "4.0.0"
 time = "0.3.26"
-
 tokio = { version = "1.39.0", features = ["full"] }
 
 [features]
